# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Vakt API
  version: "0.23.0"
  description: |
    Self-hosted Security & Compliance Platform API.
    All data stays within your infrastructure — no phone-home, no cloud dependency.

    **Authentication:** Paseto v2 bearer tokens. Obtain a token via `POST /auth/login`.
    Include it as `Authorization: Bearer <token>` on all protected endpoints.

    **Base URL:** `/api/v1`

    **Pagination:** `?page=1&limit=25` on list endpoints.

    **Error format:**
    ```json
    { "error": "message", "code": "ERROR_CODE", "details": {} }
    ```
servers:
  - url: /api/v1
    description: Vakt API v1
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Paseto
...
  schemas:
...
    DORAThirdParty:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        service_type:
          type: string
        criticality:
          type: string
          enum: [low, medium, high, critical]
        country:
          type: string
        contact_email:
          type: string
          format: email
        notes:
          type: string
        linked_controls_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Authority:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        country:
          type: string
        type:
          type: string
          enum: [data_protection, cybersecurity, sector_specific, financial]
        notification_email:
          type: string
          format: email
        notification_url:
          type: string
          format: uri
        deadline_hours:
          type: integer
          description: Reporting deadline in hours (e.g. 72 for GDPR Art. 33)
        applicable_frameworks:
          type: array
          items:
            type: string
...
paths:
...
  /vaktcomply/dora/third-parties:
    get:
      summary: List DORA IKT third-party providers
      operationId: listDORAThirdParties
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of third-party providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DORAThirdParty'
    post:
      summary: Register a DORA IKT third-party provider
      operationId: createDORAThirdParty
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, service_type, criticality]
              properties:
                name: { type: string }
                service_type: { type: string }
                criticality: { type: string, enum: [low, medium, high, critical] }
                country: { type: string }
                contact_email: { type: string, format: email }
                notes: { type: string }
      responses:
        '201':
          description: Third-party provider created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DORAThirdParty'
  /vaktcomply/dora/third-parties/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get a DORA IKT third-party provider
      operationId: getDORAThirdParty
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Third-party provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DORAThirdParty'
    patch:
      summary: Update a DORA IKT third-party provider
      operationId: updateDORAThirdParty
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                service_type: { type: string }
                criticality: { type: string, enum: [low, medium, high, critical] }
                country: { type: string }
                contact_email: { type: string, format: email }
                notes: { type: string }
      responses:
        '200':
          description: Updated third-party provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DORAThirdParty'
    delete:
      summary: Delete a DORA IKT third-party provider
      operationId: deleteDORAThirdParty
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Deleted
  /vaktcomply/dora/third-parties/{id}/controls:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Link a control to a DORA third-party provider
      operationId: linkDORAThirdPartyControl
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [control_id]
              properties:
                control_id: { type: string, format: uuid }
      responses:
        '201':
          description: Control linked
  /vaktcomply/dora/third-parties/{id}/controls/{controlId}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: controlId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      summary: Unlink a control from a DORA third-party provider
      operationId: unlinkDORAThirdPartyControl
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Unlinked

  # ── DORA Resilience Tests (Sprint 29) ────────────────────────────────────
  /vaktcomply/resilience-tests:
    get:
      summary: List DORA resilience tests
      operationId: listResilienceTests
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of resilience tests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResilienceTest'
    post:
      summary: Create a DORA resilience test
      operationId: createResilienceTest
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, test_type]
              properties:
                title: { type: string }
                test_type: { type: string, enum: [vulnerability_assessment, penetration_test, tlpt, scenario_based] }
                scheduled_date: { type: string, format: date }
                scope: { type: string }
                notes: { type: string }
      responses:
        '201':
          description: Resilience test created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResilienceTest'
  /vaktcomply/resilience-tests/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get a DORA resilience test
      operationId: getResilienceTest
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Resilience test
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResilienceTest'
    patch:
      summary: Update a DORA resilience test
      operationId: updateResilienceTest
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                status: { type: string, enum: [planned, in_progress, completed, failed] }
                scheduled_date: { type: string, format: date }
                completed_date: { type: string, format: date }
                outcome: { type: string }
                notes: { type: string }
      responses:
        '200':
          description: Updated resilience test
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResilienceTest'
    delete:
      summary: Delete a DORA resilience test
      operationId: deleteResilienceTest
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Deleted
  /vaktcomply/resilience-tests/{id}/attachment:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      summary: Upload an attachment to a resilience test
      operationId: uploadResilienceTestAttachment
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                title: { type: string }
      responses:
        '201':
          description: Attachment uploaded
...
  /vaktcomply/authorities:
    get:
      summary: List relevant regulatory authorities
      operationId: listAuthorities
      tags: [vaktcomply]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of authorities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Authority'
  /vaktcomply/org-sector:
    patch:
      summary: Update organisation sector classification
      operationId: updateOrgSector
      tags: [vaktcomply]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sector: { type: string }
                sub_sector: { type: string }
                employee_count: { type: integer }
                annual_revenue_eur: { type: number }
      responses:
        '200':
          description: Updated sector

  # ── TISAX Framework Extensions (Sprint 33–34) ────────────────────────────
...
  /vaktcomply/ai-systems/{id}/classifications:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: List risk classifications for an AI system
      operationId: listAIClassifications
      tags: [vaktcomply]
      x-pro-feature: true
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of classifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AIClassification'
...